// types.ts
export interface Post {
  id: string;
  title: string;
  excerpt: string;
  category: string;
  date: string;
  readTime: string;
  image: string;
  author: string;
  authorAvatar: string;
}

export interface Certificate {
  id: string;
  title: string;
  issuer: string;
  image: string;
  description: string;
}